Fixed lat/long bundles.

+++ app/src/main/java/nz/ac/canterbury/guessit/SinglePhotoFragment.kt
@@ -25,8 +25,8 @@ class SinglePhotoFragment : Fragment() {
         val singlePhotoImageView: ImageView = view.findViewById(R.id.singlePhotoView)
         Picasso.get().load("file:$photoPath").into(singlePhotoImageView)
 
-        val latitude = arguments?.getString("latitude")!!
-        val longitude = arguments?.getString("longitude")!!
+        val latitude = arguments?.getString("latitude")!!.toDouble()
+        val longitude = arguments?.getString("longitude")!!.toDouble()
 
         val photoDescription = imageLabeler.setPhotoDescription(BitmapFactory.decodeFile(photoPath))
